Day 11
Polymorphism
