+++ .gitignore
@@ -7,3 +7,4 @@
 /Day06_Palindrome/Palindrome/target/
 /Day08_OOPBasics/Car/target/
 /Day10_Inheritance/Vehicle/target/
+/Day11_polymorphism/Animal/target/
