+++ Day11_polymorphism/Animal/src/main/java/com/mycompany/animal/Animal.java
@@ -1,0 +1,17 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Project/Maven2/JavaApp/src/main/java/${packagePath}/${mainClassName}.java to edit this template
+ */
+
+package com.mycompany.animal;
+
+/**
+ *
+ * @author SINGER
+ */
+public class Animal {
+    public void animalSound(){
+        System.out.println("Aminals Make noises...!!!");
+    }
+    
+}
