+++ Day11_polymorphism/Animal/src/main/java/com/mycompany/animal/Dog.java
@@ -1,0 +1,17 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.mycompany.animal;
+
+/**
+ *
+ * @author SINGER
+ */
+public class Dog extends Animal{
+    @Override
+    public void animalSound(){
+        System.out.println("Dog says: Woof Woof");
+    }
+    
+}
